[FIX] IME not working properly (#674)
* parse when submit only

* fix storybook
